+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Enable join classroom button if username is populated from local storage
+- Use domain from lti jwt token when no consumer site used
 
 ## [5.2.0] - 2024-10-22